Handling different types of comment notifications
It turns out that GitHub handles notifications differently depending on _how_ the comment was submitted, and potentially even more differently if the repository has issues turned off. It's not quite clear yet exactly what the different cases are, but it appears that the location (type) of a comment varies depending on whether the comment was made on the pull request itself (issues) or on the diff section of a pull request (pulls) or on an individual commit of a pull request (commits). The solution is to parse out the type from the `subject.latest_comment_url` and get the comment from the relevant location.