Including the `dist` folder in the repository and pruning some stuff
So this really sucks but there really isn't a great way to publish an apm-specific version of the app (pre-babeled) in the same way that you can do such a thing for npm. Basically, apm wants your package to exactly match your git repository. This means there isn't really much room for transpiling in a clean way (babeling, then publishing the babeled code, excluding the unbabeled code)

See the following:
https://discuss.atom.io/t/how-to-use-the-npmignore/9208/5
https://github.com/atom/apm/issues/372